Add precision-aware dtype handling (bf16/fp16/fp32)
Enable precision-aware model construction and inference by introducing a precision argument and setting module/buffer dtypes accordingly. EoMT: add precision parsing (bf16/fp16/fp32), cast the encoder, embeddings, heads, upscale blocks, pixel mean/std and attn mask buffers to the chosen dtype, and import the FSRCNN wrapper type. FSRCNN wrapper: avoid forcing .float() on input so reduced-cast paths (e.g. bfloat16) are preserved. Tests: pass precision through build_model and switch the test autocast to bfloat16. Training: remove hardcoded float32 dtype and unnecessary .float() casts and simplify masks_to_boxes calls so tensors keep compatible dtypes. These changes make the codebase compatible with bfloat16/fp16 workflows and remove unnecessary forced float32 casts.

+++ models/scale_block.py
@@ -153,7 +153,7 @@ class FSRCNNx2YOnlyRGBWrapper(nn.Module):
         if x.ndim != 4 or x.shape[1] != 3:
             raise ValueError(f"Expected [N, 3, H, W], got {tuple(x.shape)}")
 
-        x = x.float().clamp(0.0, 1.0)
+        x = x.clamp(0.0, 1.0)
 
         ycbcr = self._rgb_to_ycbcr(x)
         y = ycbcr[:, 0:1] / 255.0       # model expects Y in [0, 1]
